+++ examples/popup.tsx
@@ -6,6 +6,7 @@ import 'rmc-picker/assets/popup.css';
 import * as React from 'react';
 import * as ReactDOM from 'react-dom';
 import PopupCascader from '../src/Popup';
+import Cascader from '../src/Cascader';
 import globalData from './data';
 import arrayTreeFilter from 'array-tree-filter';
 
@@ -17,11 +18,7 @@ const Demo = React.createClass({
       visible: false,
     };
   },
-  
-  onPickerChange(value) {
-    console.log('onPickerChange ', value);
-  },
-  
+
   onChange(value) {
     console.log('onChange', value);
     this.setState({
@@ -29,14 +26,18 @@ const Demo = React.createClass({
       visible: false,
     });
   },
-  
+
   onDismiss() {
     console.log('onDismiss');
     this.setState({
       visible: false,
     });
   },
-  
+
+  onPickerChange(value) {
+    console.log('picker change', value);
+  },
+
   getSel() {
     const value = this.state.value;
     if (!value) {
@@ -49,23 +50,28 @@ const Demo = React.createClass({
       return v.label;
     }).join(',');
   },
-  
+
   outerCtrl() {
     this.setState({
       visible: !this.state.visible,
     });
   },
-  
+
   render() {
+    const cascader = (
+      <Cascader
+        rootNativeProps={{'data-xx':'yy'}}
+        onChange={this.onPickerChange}
+        data={globalData}
+        cols={COLS}
+      />
+    );
     return (<div style={{ padding: 10 }}>
       <h3>popup cascader</h3>
       <p>选择的城市：{this.getSel()}</p>
       <PopupCascader
-        pickerRootNativeProps={{'data-xx':'yy'}}
-        data={globalData}
+        cascader={cascader}
         value={this.state.value}
-        cols={COLS}
-        onPickerChange={this.onPickerChange}
         onDismiss={this.onDismiss}
         onChange={this.onChange}
         title="Cascader"
@@ -75,11 +81,9 @@ const Demo = React.createClass({
       <h3>just cascader no children</h3>
       <button onClick={this.outerCtrl}>switch</button>
       <PopupCascader
+        cascader={cascader}
         visible={this.state.visible}
-        data={globalData}
         value={this.state.value}
-        cols={COLS}
-        onPickerChange={this.onPickerChange}
         onDismiss={this.onDismiss}
         onChange={this.onChange}
       />