+++ HISTORY.md
@@ -1,9 +1,14 @@
 # History
 ----
 
+## 3.0.0 / 2016-09-07
+
+- extract cascader from popup
+
+
 ## 2.0.0 / 2016-06-16
 
-- support react-native and typescript`
+- support react-native and typescript
 
 ## 0.2.5 / 2015-01-18